requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.4.0
asgiref==3.11.0
asn1crypto==1.5.1
asttokens==3.0.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.4.0
Babel==2.17.1
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.6
charset-normalizer==3.6.0
colorama==0.4.6
cryptography==45.0.2
decorator==5.1.1
Django==5.2.11
django-cors-headers==4.6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
executing==2.3.1
Flask==3.1.1
frozenlist==1.6.0
greenlet==4.1.1
idna==3.10
ipython==8.34.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
lxml==6.0.2
magic-filter==1.0.12
MarkupSafe==3.0.3
marshmallow==4.1.2
matplotlib-inline==0.2.1
multidict==6.7.0
openpyxl==3.1.5
parso==0.8.5
pika==1.3.2
pillow==12.0.0
plum-dispatch==1.7.4
prompt_toolkit==3.0.52
propcache==0.4.1
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
Pygments==2.19.2
PyJWT==2.11.0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
qrcode==8.2
requests==2.32.5
rsa==4.9.1
soupsieve==2.6.1
spire-doc==13.8.0
sqlparse==0.5.5
stack-data==0.6.3
Telethon==1.40.0
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
wcwidth==0.2.14
Werkzeug==3.1.4
yarl==1.22.0
